refactor: move deserializer tests to AliasAwareIndexResponseDeserializerTest
Deserialization unit tests belong with the class they test, not in
ElasticsearchDataStoreClientTest which should focus on client behaviour.
